+++ tools/visual-pipeline-and-platform-evaluation-tool/ui/nginx.conf
@@ -170,8 +170,8 @@ server {
         application/xml+rss
         application/json;
 
-    # Cache static assets (exclude /assets/images and /assets/videos)
-    location ~* ^/(?!assets/(images|videos)).*\.(js|css|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf|eot)$ {
+    # Cache static assets
+    location ~* \.(js|css|png|jpg|jpeg|gif|ico|svg|woff|woff2|ttf|eot)$ {
         expires 1y;
         add_header Cache-Control "public, immutable";
     }
