+++ tools/visual-pipeline-and-platform-evaluation-tool/ui/src/lib/fileUtils.ts
@@ -36,3 +36,16 @@ export const getFilenameFromPath = (value: unknown): string => {
       .pop() ?? ""
   );
 };
+
+/**
+ * Formats a byte size into a human-readable string
+ * @param bytes - The size in bytes
+ * @returns A formatted string with the appropriate unit (B, KB, MB, GB)
+ */
+export const formatBytes = (bytes: number): string => {
+  if (bytes === 0) return "0 B";
+  const k = 1024;
+  const sizes = ["B", "KB", "MB", "GB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return `${(bytes / Math.pow(k, i)).toFixed(2)} ${sizes[i]}`;
+};
